restructure to support browser
